StudymateX/requirements.txt
streamlit>=1.24.0
pymupdf>=1.22.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
numpy>=1.24
tqdm
python-dotenv
groq>=0.5.0
transformers>=4.35.0
torch>=2.0.0
huggingface-hub>=0.19.0
accelerate>=0.24.0
